[Style] User 디렉토리 재구성 - #13
