--- favor/src/main/java/com/favor/favor/User/SignInDto.java
@@ -1,4 +1,0 @@
-package com.favor.favor.user;
-
-public class SignInDto {
-}
