+++ favor/src/main/java/com/favor/favor/anniversary/AnniversaryService.java
@@ -1,13 +1,12 @@
 package com.favor.favor.anniversary;
 
 import com.favor.favor.exception.CustomException;
-import com.favor.favor.user.User;
-import com.favor.favor.user.UserRepository;
+import com.favor.favor.member.User;
+import com.favor.favor.member.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
