+++ favor/src/main/java/com/favor/favor/gift/GiftRequestDto.java
@@ -2,7 +2,7 @@ package com.favor.favor.gift;
 
 import com.favor.favor.common.enums.Category;
 import com.favor.favor.common.enums.Emotion;
-import com.favor.favor.user.User;
+import com.favor.favor.member.User;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
