+++ favor/src/main/java/com/favor/favor/member/SignInDto.java
@@ -1,0 +1,4 @@
+package com.favor.favor.member;
+
+public class SignInDto {
+}
